Challenge 2:  re-arranging ordering, making full use of type information.
Turns out to be an important thing to specify the particular type
parameter you want to use in the invocation of a function which
uses a type parameter, and it"s important, sometimes, to specify
that a generic type parameter is a "subtype" of, "extension" of,
a more specific type.  The type parameter "T extends todoItem" is
what allowed the compiler to conclude that there is an "id" property
available for the constituents of the items array.