Notes added:  stream-of-consciousness thoughts
Thoughts added while watching instructor' presentation of Ch03, video 5.
This is the "initial version" before derivation of the final TypeScript
program in the "03_05e" folder.  So, he addressed the issues identified,
and actually showed how the TypeScript compiler helps to identify the
kinds of questions I raised in the notes I had addded, although the
issue is dealt with in a slightly different fashion, in the end.